% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characters.R
\name{get_characters}
\alias{get_characters}
\title{Get Characters by Character Names}
\usage{
get_characters(name)
}
\arguments{
\item{name}{The name of the character. No defalut value, and cannot be left blank or error occurs}
}
\description{
Use this function to search for specific Marvel character from Marvel API with character' name
}
\details{
The function askes the user to input the name of a specific Marvel character, if the character
exists, this function returns information of characters ID, name, description, comics,series,
events and stories information.
}
\keyword{Character}
\keyword{Character,}
\keyword{Comics,}
\keyword{Marvel,}
\keyword{name}
