% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comics.R
\name{get_comics}
\alias{get_comics}
\title{Get Comics by Comics Title}
\usage{
get_comics(title)
}
\arguments{
\item{title}{The title of the comics. No defalut value, and cannot be left blank or error occurs}
}
\description{
Use this function to search for specific comics from Marvel API with comics titles
}
\details{
The function askes the user to input the title or keyword of a particular comics, if the related
information exists,this function returns information of comics ID, title, description, series,
creators, characters, stories and events information. Otherwise, there will be error messages
with instructions to correct.
}
\keyword{Comics}
\keyword{Comics,}
\keyword{Marvel,}
\keyword{Title}
