% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Get Events by Event Names}
\usage{
get_events(name)
}
\arguments{
\item{name}{The name of the event. No defalut value, and cannot be left blank or error occurs}
}
\description{
Use this function to search for specific Marvel event from Marvel API with event's name
}
\details{
The function askes the user to input the name of a particular Marvel event, if the event
exists, this function returns information of events ID, name, description, comics, characters,
series, creators and stories information.
}
\keyword{Comics,}
\keyword{Event}
\keyword{Events,}
\keyword{Marvel,}
\keyword{name}
