% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_series}
\alias{get_series}
\title{Get Series by Series Titles}
\usage{
get_series(title)
}
\arguments{
\item{name}{The title of the series. No defalut value, and cannot be left blank or error occurs}
}
\description{
Use this function to search for specific Marvel series from Marvel API with series' title
}
\details{
The function askes the user to input the title of a particular Marvel series, if the series
exists, this function returns information of series ID, name, description, comics, creators,
series, events and stories information.
}
\keyword{Comics,}
\keyword{Marvel,}
\keyword{Series}
\keyword{Series,}
\keyword{title}
