% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_characters.R
\name{search_characters}
\alias{search_characters}
\title{Search characters from Marvel}
\usage{
search_characters(id = "", type = "")
}
\arguments{
\item{id}{Each character's unique ID number. The default is NULL so without input,
it will reutrn random characters for the number of limits set in token_to_query().}

\item{type}{Type of Information from Marvel, Default is Null, choose one from
comics, events, series, and stories.}
}
\description{
Use this function to search for characters from Marvel API
}
\details{
The function retrieves characters of Marvels from the user's input on characters id number
and/or the type of information prefered, including comics, events, series, and stories.
Please note, creators cannot be retireved using this function.
}
\keyword{Comics,}
\keyword{Marvel,}
\keyword{characters}
