% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_comics.R
\name{search_comics}
\alias{search_comics}
\title{Search comics from Marvel}
\usage{
search_comics(id = "", type = "")
}
\arguments{
\item{id}{Each comics' unique ID number. The default is Null. Without input,
it will reutrn random comics with the count limit set in tooken_to_query().}

\item{type}{Type of Information from Marvel, Default is Null, choose one from
creators, stories, events, and characters.}
}
\description{
Use this function to search for comics from Marvel API
}
\details{
The function retrieves comics of Marvels from the user's input on comics id number
and/or the type of information prefered, including creators, stories, events, and characters.
Please note, series cannot be tetireved using this function.
}
\keyword{Comics,}
\keyword{Marvel,}
\keyword{comics}
