% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_creators.R
\name{search_creators}
\alias{search_creators}
\title{Search creators from Marvel}
\usage{
search_creators(id = "", type = "")
}
\arguments{
\item{id}{Each creator's unique ID number. The default is Null. Without input,
it will reutrn random creators with the count limit set in tooken_to_query().}

\item{type}{Type of Information from Marvel, Default is Null, choose one from
comics, events, series and stories.}
}
\description{
Use this function to search for creators from Marvel API
}
\details{
The function retrieves creators of Marvels from the user's input on creators id number
and/or the type of information prefered, including comics, stories, series, and events.
Please note, characters cannot be retireved using this function.
}
\keyword{Comics,}
\keyword{Marvel,}
\keyword{creators}
