% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_events.R
\name{search_events}
\alias{search_events}
\title{Search events from Marvel}
\usage{
search_events(id = "", type = "")
}
\arguments{
\item{id}{Each event's unique ID number. The default is Null. Without input,
it will reutrn random events with the count limit set in tooken_to_query().}

\item{type}{Type of Information from Marvel, Default is Null, choose one from
comics, stories, creators, characters, and series.}
}
\description{
Use this function to search for events from Marvel API
}
\details{
The function retrieves events of Marvels from the user's input on events id number
and/or the type of information prefered, including comics, stories, creators, characters,
and series.
}
\keyword{Comics,}
\keyword{Marvel,}
\keyword{events}
