% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_series.R
\name{search_series}
\alias{search_series}
\title{Search series from Marvel}
\usage{
search_series(id = "", type = "")
}
\arguments{
\item{id}{Each series' unique ID number. The default is Null. Without input,
it will reutrn random seires with the count limit set in tooken_to_query().}

\item{type}{Type of Information from Marvel, Default is Null, choose one from
comics, events, creators, characters, and stories.}
}
\description{
Use this function to search for series from Marvel API
}
\details{
The function retrieves series of Marvels from the user's input on series id number
and/or the type of information prefered, including comics, events, creators, characters,
and stories.
}
\keyword{Comics,}
\keyword{Marvel,}
\keyword{series}
