% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_stories.R
\name{search_stories}
\alias{search_stories}
\title{Search stories from Marvel}
\usage{
search_stories(id = "", type = "")
}
\arguments{
\item{id}{Each story's unique ID number. The default is Null. Without input,
it will reutrn random stories with the count limit set in tooken_to_query().}

\item{type}{Type of Information from Marvel, Default is Null, choose one from
comics, events, creators, characters, and series.}
}
\description{
Use this function to search for stories from Marvel API
}
\details{
The function retrieves stories of Marvels from the user's input on stories id number
and/or the type of information prefered, including comics, events, creators, characters,
and series.
}
\keyword{Comics,}
\keyword{Marvel,}
\keyword{stories}
